Init
I removed all of my sensative information (API keys/phone number/etc) in this, and you can see a lot of missing variables. Of course, I coded this within 24 hours at hackupstate.org, so this version isn't anything special. It assumes that all of the files are located in /home. All of the ugly stuff is going to change. Soon.